Small refactor of code to use JS coding standards. Also separating Postgres interfaces into different files.
