+++ api/routes/users.js
@@ -4,18 +4,13 @@ var router = express.Router();
 const {
   getUsers,
   getUserById
-  // testFunc
-} = require('../db/connect.js');
+} = require('../db/db-user-interface.js');
 
 /* GET users listing. */
 // router.get('/', function(req, res, next) {
 //   res.send('respond with a resource');
 // });
 
-// router.route('/').get(function(req, res, next) {
-//     res.send('respond with a resource');
-//   });
-// router.route('/').get(testFunc);
 router.route('/').get(getUsers);
 router.route('/:id').get(getUserById);
